pyproject.toml
[tool.poetry]
name = "warcraft-raid-organizer"
version = "0.1.0"
description = "World of Warcraft raid organization project"
authors = ["Ryan Davis <ryancraigdavis@github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
quart = "^0.19.6"
databases = "^0.9.0"
asyncpg = "^0.29.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
